requirements.txt
feedparser
pandas
numpy
sentence-transformers
scikit-learn
beautifulsoup4
python-dateutil
google-generativeai
google-cloud-storage
python-dotenv
